moved is wihtin note to stillWithinNote and added isWithinNote placeholders
